refactor(core): 按域拆分 db schema 文件
