+++ noj-core/src/db/schema/index.ts
@@ -1,0 +1,9 @@
+export * from "./identity.ts";
+export * from "./catalog.ts";
+export * from "./objective.ts";
+export * from "./gateway.ts";
+export * from "./system.ts";
+export * from "./contest.ts";
+export * from "./submission.ts";
+export * from "./messaging.ts";
+export * from "./community.ts";
